Checkpoint

+++ bc-excuse-to-learn.py
@@ -8,20 +8,35 @@ import cloud;
 import os;
 
 # do all work in temporary (but fixed) directory
-exists = os.path.isdir("/tmp/bcetlp")
-if not(exists): os.mkdir("/tmp/bcetlp")
-os.chdir("/tmp/bcetlp")
+tmpdir = "/tmp/bcetlp"
+if not(os.path.isdir(tmpdir)): os.mkdir(tmpdir)
+os.chdir(tmpdir)
+
+print tmpdir+"/metar.txt"
 
 # get both files I need, in parallel (to save time)
-parfile = open("parallel", "w")
-parfile.write("curl http://weather.aero/dataserver_current/cache/metars.cache.csv.gz | gunzip | tail -n +6 > metar.txt\n")
-parfile.write("curl -o buoy.txt http://www.ndbc.noaa.gov/data/latest_obs/latest_obs.txt\n")
-parfile.close()
+test = not(os.path.isfile(tmpdir+"/metar.txt") and
+        os.path.isfile(tmpdir+"/buoy.txt"))
 
-print "ALPHA"
+print test
 
-os.system("ls -l")
-print os.system("cat parallel 1> output")
-os.system("/usr/local/bin/parallel < parallel 1> par.out 2> par.err")
+exit(0)
 
+"""
+ if (not(and(os.path.isfile("$tmpdir/metar.txt"), os.path.isfile("$tmpdir/buoy.txt")))):
+       print "Start"
+       parfile = open("parallel", "w")
+       parfile.write("curl http://weather.aero/dataserver_current/cache/metars.cache.csv.gz | gunzip | tail -n +6 > metar.txt\n")
+       parfile.write("curl -o buoy.txt http://www.ndbc.noaa.gov/data/latest_obs/latest_obs.txt\n")
+       parfile.close()
+       os.system("/usr/local/bin/parallel < parallel 1> par.out 2> par.err")
+
+
+print "Done"
+
+"""
+
+       
+
+
